+++ src/main/java/org/ingestor/DocGenerator.java
@@ -4,5 +4,5 @@ import com.couchbase.client.java.json.JsonObject;
 
 public interface DocGenerator {
 
-    JsonObject generateDoc(int counter);
+    JsonObject generateDoc(long counter);
 }
